+++ bookcut/mirror_checker.py
@@ -1,10 +1,19 @@
 import requests
+import configparser
+import logging
+import os
 
 
 def main():
+    '''Check which LibGen mirror is available'''
 
-    mirrors = ['https://libgen.lc/', 'http://libgen.li/',
-               'http://185.39.10.101/', 'http://genesis.lib/']
+    config = configparser.ConfigParser()
+    module_path = os.path.dirname(os.path.realpath(__file__))
+    settings_ini = os.path.join(module_path, 'Settings.ini')
+    config.read(settings_ini)
+    mirrors = config.get("LibGen", "mirrors")
+    mirrors = mirrors.split(',')
+
     for url in mirrors:
         try:
             r = requests.head(url)